refactor(#97): vehicles/components — bulletproof-react segmentation
git mv:
- NewVehicleFab.tsx, Save{Error,Success}Snackbar.tsx → components/
- VehicleDetails{,Skeleton}.tsx, VehicleEditForm.tsx, VehicleView.tsx → components/
- projection/vehicleViewConfig.tsx → components/

Cross-area references back to root files (VehicleEditForm, VehicleDetails)
in stores/, utils/, api/, hooks/ updated to ../components/. External call
sites updated: App.tsx (VehicleView), VehicleCreatePage (VehicleEditForm,
Save{Error,Success}Snackbar).

Empty projection/ and xml/ directories removed.
