+++ src/data/vehicles/components/VehicleDetailsSkeleton.tsx
