+++ src/data/vehicles/components/VehicleEditForm.tsx
@@ -1,11 +1,11 @@
 import { TextField, Typography, Divider } from '@mui/material';
 import { useTranslation } from 'react-i18next';
-import type { Vehicle } from './types/Vehicle';
-import type { VehicleModel } from './types/VehicleModel';
-import { firstText } from '../netex/multilingualString.ts';
-import { MISSING_TEXT } from './utils/vehicleFormDefaults';
-import { intToRef, refToInt } from './utils/transportTypeRef';
-import { FormLayout, FieldRow } from '../../components/FormLayout.tsx';
+import type { Vehicle } from '../types/Vehicle';
+import type { VehicleModel } from '../types/VehicleModel';
+import { firstText } from '../../netex/multilingualString.ts';
+import { MISSING_TEXT } from '../utils/vehicleFormDefaults';
+import { intToRef, refToInt } from '../utils/transportTypeRef';
+import { FormLayout, FieldRow } from '../../../components/FormLayout.tsx';
 
 export interface VehicleEditFormValue {
   vehicle: Vehicle;
