+++ src/data/vehicles/components/vehicleViewConfig.tsx
