+++ src/data/vehicles/stores/vehicleFormState.ts
@@ -5,7 +5,7 @@
  * frame (M2, PR #74 review). All edits go through `edit`, which leaves
  * the baseline untouched so divergence is the only signal for dirty.
  */
-import type { VehicleEditFormValue } from '../VehicleEditForm';
+import type { VehicleEditFormValue } from '../components/VehicleEditForm';
 import type { Vehicle } from '../types/Vehicle';
 import { BLANK_FORM, MISSING_MODEL } from '../utils/vehicleFormDefaults';
 import { TRANSPORT_TYPE_REF_PATTERN } from '../utils/transportTypeRef';
