+++ src/data/vehicles/utils/vehicleFormDefaults.ts
@@ -1,5 +1,5 @@
 import type { VehicleModel } from '../types/VehicleModel';
-import type { VehicleEditFormValue } from '../VehicleEditForm';
+import type { VehicleEditFormValue } from '../components/VehicleEditForm';
 
 export const INLINE_MODEL_ID = 'INLINE:VehicleModel:1' as const;
